+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## v1.1.10 (1.20.1)
 - fixed issue where miapi would break certain tags
+- made LoreProperty usable
 - changed module property resolving to allow for more complex conditions
 - added HideFlags property